+++ src/defs.hpp
@@ -6,8 +6,8 @@
 #include <unordered_map>
 
 // Global screen dimensions, updated on window resize
-static size_t SCREEN_WIDTH = 800;
-static size_t SCREEN_HEIGHT = 600;
+static int SCREEN_WIDTH = 800;
+static int SCREEN_HEIGHT = 600;
 
 constexpr float EPSILON_0 = 8.8541878128e-12f;
 constexpr float K_E = 1 / (4 * std::numbers::pi_v<float> * EPSILON_0);
@@ -16,7 +16,9 @@ constexpr float GLOBAL_SCALE = 100.0f;
 constexpr int FONT_SIZE = 24;
 
 template <> struct std::formatter<raylib::Vector2> {
-  constexpr auto parse(std::format_parse_context const &ctx) const { return ctx.begin(); }
+  constexpr auto parse(std::format_parse_context const &ctx) const {
+    return ctx.begin();
+  }
   template <typename FormatContext>
   auto format(const raylib::Vector2 &c, FormatContext &ctx) const {
     return std::format_to(ctx.out(), "({}, {})", c.x, c.y);
@@ -24,7 +26,9 @@ template <> struct std::formatter<raylib::Vector2> {
 };
 
 template <> struct std::formatter<raylib::Color> {
-  constexpr auto parse(std::format_parse_context const &ctx) const { return ctx.begin(); }
+  constexpr auto parse(std::format_parse_context const &ctx) const {
+    return ctx.begin();
+  }
   template <typename FormatContext>
   auto format(const raylib::Color &c, FormatContext &ctx) const {
     return std::format_to(ctx.out(), "({}, {}, {}, {})", c.r, c.g, c.b, c.a);
